+++ bin/langs.sh
@@ -1,0 +1,2 @@
+#!/usr/bin/env bash
+php cgtest.php --config=.cgtest.full.php --lang-versions --quiet "$@"